+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Lotus::Model
 A persistence layer for Lotus
 
+## v0.5.2 - 2016-01-19
+### Changed
+- [Sean Collins] Improved error message for `Lotus::Model::Adapters::NoAdapterError`
+
+### Fixed
+- [Kyle Chong & Trung Lê] Catch Sequel exceptions and re-raise as `Lotus::Model::Error`
+
 ## v0.5.1 - 2016-01-12
 ### Added
 - [Taylor Finnell] Let `Lotus::Model::Configuration#adapter` to accept arbitrary options (eg. `adapter type: :sql, uri: 'jdbc:...', after_connect: Proc.new { |connection| connection.auto_commit(true) }`)